Let camera_wrappers build without dependencies.
camera_wrappers can now build without Kinect1X and OpenNI2 libraries (in
case you want to just process serialized streams).

A few tiny updates for C++11 now that we are building with VS2015.

+++ apps/rgbd_recorder/src/Viewfinder.cpp
@@ -48,7 +48,7 @@ Viewfinder::Viewfinder( const std::vector< StreamConfig >& streamConfig,
 {
     if( m_oniCamera->isValid() )
     {
-        for( size_t i = 0; i < m_streamConfig.size(); ++i )
+        for( int i = 0; i < static_cast< int >( m_streamConfig.size() ); ++i )
         {
             const StreamConfig& config = m_streamConfig[ i ];
             m_outputMetadata.push_back( StreamMetadata{
